+++ src/app/login/page.tsx
@@ -1,8 +1,8 @@
 import LoginButton from '@/components/auth/LoginButton';
-import styles from './page.module.css';
 import PayPalDonate from '@/components/auth/PayPalDonate';
-import Header from '@/components/header/Header';
-import Footer from '@/components/Footer';
+import Header from '@/components/layout/Header';
+import Footer from '@/components/layout/Footer';
+import styles from './page.module.css';
 
 export default function LoginPage() {
   return (
