+++ src/app/room/[id]/page.tsx
@@ -1,11 +1,10 @@
 'use client';
 
 import MyRooms from '@/components/room/sidebar/MyRooms';
-import UserProfile from '@/components/header/profile/UserProfileButton';
+import UserProfile from '@/components/profile/UserProfileButton';
 import ToggleSidebar from '@/components/room/sidebar/ToggleSidebar';
 import SettingMap from '@/components/room/place/SettingMap';
 import Sidebar from '@/components/room/sidebar/Sidebar';
-import { useQueryUser } from '@/hooks/useQueryUser';
 import { useEffect, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 import { useQuery } from '@tanstack/react-query';
@@ -15,12 +14,11 @@ import styles from './page.module.css';
 
 const SettingPage = () => {
   const router = useRouter();
-  const { id: userId } = useQueryUser();
   const { id: roomId }: { id: string } = useParams();
   const [isOpened, setIsOpened] = useState(true);
 
   const { data: room } = useQuery({
-    queryKey: ['room', 'confirmed', 'createdBy'],
+    queryKey: ['room', 'confirmed', 'createdBy', roomId],
     queryFn: () => getRoomIsConfirmed(roomId),
     select: (data) => data[0]
   });
